Moved shared tester_dataclasses to a separate package.
This way other packages can freely access these shared dataclasses (Algorithm, Parser...) without triggering plugins lookup.
