+++ analysis/tester_results/TEST/tmp.sh
@@ -1,0 +1,10 @@
+#!/bin/bash
+
+tmp_file=".tmp"
+
+for file in *.dat;
+do
+	sort -k1 -V $file > $tmp_file
+	rm $file
+	mv $tmp_file $file
+done
