+++ package_algorithms/dp.py
@@ -1,6 +1,6 @@
 from typing import List, Dict
 import numpy as np
-from algorithm_tester.tester_dataclasses import Algorithm, AlgTesterContext
+from algorithm_tester_common.tester_dataclasses import Algorithm, AlgTesterContext
 from package_algorithms.alg_dataclasses import ConfigCounter, TaskKnapsackProblem, Thing, RecursiveResult, Solution, base_columns
 
 class DynamicProgramming_Weight(Algorithm):
